POMP.jl

Summary.

Package Features

Implement POMP models.

Simulate models with process noise and measurement error.

Particle filters.

Function Documentation

Basic constructor

POMP.pomp — Function

pomp is the constructor for the PompObject class.

pomp(
    data;
    t0, times,
    accumvars,
    rinit, rprocess,
    rmeasure, logdmeasure
    )

Arguments

data: observations. The default constructor takes a vector of NamedTuples as data. One can also supply a DataFrame.

t0: zero time, t₀.

times: observation times. If data is supplied as a DataFrame, times should be a Symbol which is the time variable in the DataFrame.

accumvars: a NamedTuple of state variables to be reset (usually to zero) immediately following each observation.

rinit: simulator of the latent-state distribution at t₀.

rprocess: simulator of the latent-state process.

rmeasure: simulator of the measurement process.

logdmeasure: log pdf of the measurement process.

Alternatively, one can construct a PompObject from a DataFrame. In this case, times should be the Symbol of the time-variable.

Given an AbstractPompObject, object, pomp(object) returns the underlying concrete PompObject. Calling pomp(object, args...) returns a copy of object, modified according to args....

POMP.pomp! — Function

pomp!(object, args...) modifies the PompObject object in place. Individual basic components can be modified, set, or unset.

Simulation

POMP.simulate — Function

simulate(object; params, nsim = 1, args...)

simulate simulates the POMP. args... can be used to modify or unset fields.

POMP.simulate! — Function

simulate!(object; args...)

simulate! simulates in place. args... can be used to modify or unset fields.

Particle filter

POMP.pfilter — Function

pfilter(object; params, Np = 1, args...)

pfilter runs a basic particle filter. args... can be used to modify or unset fields.

Workhorses

POMP.rinit — Function

rinit(object; t0=timezero(object), params=coef(object), nsim=1)

rinit is the workhorse for the simulator of the initial-state distribution.

The user can supply an rinit component as a function that takes parameters and, optionally, t0, the initial time.

Arguments

object: the PompObject

params: a NamedTuple of parameters

t0: the time at which rinit is to be simulated. This should be a single scalar.

nsim: the number of simulations desired.

POMP.rmeasure — Function

rmeasure(object; x, times=times(object), params=coef(object))

rmeasure is the workhorse for the simulator of the measurement distribution.

The user can supply an rmeasure component as a function that takes states, parameters, and, optionally, t, the current time.

POMP.rprocess! — Function

rprocess(object; x0, t0 = timezero(object), times=times(object), params=coef(object))

rprocess is the workhorse for the simulator of the process

The user can supply an rprocess component as a function that takes states, parameters, and current time (t) and returns the updated time and state.

If there is no user-supplied rprocess component, the dynamics are trivial.

POMP.logdmeasure! — Function

logdmeasure(object; times=times(object), y = obs(object), x = states(object), params)

logdmeasure is the workhorse for the evaluator of the log measurement density.

The user can supply a logdmeasure component as a function that takes data, states, parameters, and, optionally, t, the current time.

Helper functions

POMP.coef — Function

coef(object::SimPompObject)

coef extracts the parameters stored in a SimPompObject.

coef(object::PfilterdPompObject)

coef extracts the parameters stored in a PfilterdPompObject.

POMP.states — Function

states(object::AbstractSimPompObject)

states extracts the state trajectory of a SimPompObject.

POMP.obs — Function

obs(object)

obs extracts the time vector of a PompObject.

obs(object::AbstractSimPompObject)

obs extracts the observations of a SimPompObject.

POMP.times — Function

times(object)

times extracts the time vector of a PompObject.

POMP.timezero — Function

timezero(object)

timezero extracts the zero-time (t0) of a PompObject.

Examples

POMP.sir — Function

sir(
    β = 0.5, γ = 0.25, N = 10000,
    ρ = 0.3, k = 10,
    S₀ = 0.9, I₀ = 0.01, R₀ = 0.1,
    δt = 0.1, t₀ = 0.0,
    times = [t for t ∈ range(start=1.0,stop=90,step=1.0)]
   )

sir returns a SimPompObject containing simulated SIR data.

Parameters

β: transmission rate

γ: recovery rate

N: population size

ρ: reporting rate

k: overdispersion coefficient (negative binomial size parameter)

S₀, I₀, R₀: relative proportions of susceptible, infected, recovered (respectively) in the population at t=t₀.

δt: Euler stepsize

t₀: zero-time

times: vector of observation times

POMP.gompertz — Function

gompertz()

gompertz is a PompObject containing Parus major data and a simple Gompertz population model. The population model has a single scalar state variable, $X_t$, which obeys

\[X_t = X_{t-1}^S\,K^{1-S}\,\varepsilon_t,\]

where $S = e^{-r\delta{t}}$ and $\varepsilon_t \sim \mathrm{LogNormal}(0,\sigma_p)$. The time-step is one unit: $\delta{t}=1$. The data are assumed to be drawn from a log-normal distribution. In particular,

\[Y_t \sim \mathrm{LogNormal}(\log{X_t},\sigma_m).\]

Parameters

r: the growth rate

K: the equilibrium population density

X₀: the initial population density

σₚ: process noise s.d.

σₘ: measurement noise s.d.